Make frontend less self-reliant and removed some of its attributes. If it needs them, it can ask the backend.
The frontend for now will have its own gsettings object. The idea is that gsettings is the glue between the front and backend; front read/writes to gsettings, backend reads and then does logic based on that.

Added more getters to backend to allow frontend or anything to grab values they need without storing them on their own.

Cleaned up some formatting and other things.

This app is still a complete mess. Lots to do.