Extracted backend constructor out of window into main but it'll likely be moved into the backend module itself.
Temporarily removed the reset logic. Will reimplement it later with a cleaner design.

Temporarily shelved the dynamic AppDetails and replaced it with a specialized one for Firefox. This is to help solidify the downstream implementations so that it's clearer what attributes AppDetails needs to hold. Once it's settled, I'll refactor to make it dynamic again. But as of now, the dynamism is causing headaches and too much shifting around.

Moved autofind paths schema key to be app specific.