online: improve & clean up 'get_updates_online()'
Getting updates now handles if the files are missing but no
update is available. The prior workaround was to reset the app.
This method is also a much cleaner now.

`installed_version` is now set when the files are downloaded
to prevent edge case where app may redundantly re-download
the files if the user hasn't manually enabled the theme since
an update.