Improved Online & Install Managers (#7)
install: Install requests are now routed through a single method and an optional argument is available to choose whether you want a fast or quick install. This reduces a good amount of duplicate code.

online: Improved getting updates to allow for re-downloading theme files if lost — even if there is no update available. Now the "installed_version" is updated when an update is detected as available, not when it's installed.

general: Thoroughly cleaned code based on lint suggestions and fixed a lot of type hints issues (mypy)