Changing package format refreshes the GUI page
This removes the need to restart the app.

The only issue is this freezes the app for a second while the page gets
reconstructed, which sucks. In the future, it could be changed to only
refresh the profiles but that would require a bigger redesign of the
page.

I'd like to add more features like Thunderbird support first.